Use -apply-system font on error page (#140)
The CSS rule on the error page doesn't work and the text on the page isn't rendered in SF UI Text. If we want to use SF on the error page, then, at least in Safari, we can use `-apply-system` as the font family.

I also added a fallback to sans-serif, just to be explicit.