Fix: with-passport example dependency issue (#24567)
fixes #24401

Simply bumping `swr` to the latest version seems to fix the issue. 🥂 

Couldn't find a consistent pattern across the examples, but most examples are flexible with swr version and use the semver caret.

Very standard use of the useSWR hook is there in the example with nothing obviously version specific.