docs(examples/with-redux-persist): add 'Open in StackBlitz' button (#25972)
Add `Open in StackBlitz` button to `examples/with-redux-persist` and `examples/with-videojs`. 

#### with-redux-persist
At first we though the example didn't work correctly in WebContainer because it shows the following message

> redux-persist failed to create sync storage. falling back to noop storage.

After investigating this, we saw the same message on local as well and is probably intended behaviour because of SSR. The example works, but it makes sense that it can't get the local storage on the server side and that's what causing the message.

#### with-videojs
The video's couldn't be played in the preview frame before. I fixed it on our end and thus also added the StackBlitz button.

## Bug

- [ ] Related issues linked using `fixes #number`
- [ ] Integration tests added

## Feature

- [ ] Implements an existing feature request or RFC. Make sure the feature request has been accepted for implementation before opening a PR.
- [ ] Related issues linked using `fixes #number`
- [ ] Integration tests added
- [ ] Documentation added
- [x] Example(s) updated
- [ ] Telemetry added. In case of a feature if it's used or not.

## Documentation / Examples

- [x] Make sure the linting passes