Update README of cms-cosmic example (#31701)
## Documentation / Examples - Update COSMIC_BUCKET_SLUG location in cms-cosmic example
While trying the cms-cosmic example I realized that COSMIC_BUCKET_SLUG  is actually in the **API Access** menu and it's not in the **Basic Settings** menu.
### Screenshots
Here are some screenshots from the two sections:
#### Basic Settings
![basic_settings](https://user-images.githubusercontent.com/16997807/142915282-49fb8321-a01e-4478-9047-29274da36852.png)
#### API Access
![API_access](https://user-images.githubusercontent.com/16997807/142915289-74967622-adbb-405c-a1c8-5c02c2249675.png)

### Testing
I've tried the example using the COSMIC_BUCKET_SLUG from API Access and it worked without issues.